+++ src/config.js
@@ -1,0 +1,6 @@
+'use strict'
+
+module.exports = {
+  PROTOCOL: '/ipfs/ping/1.0.0',
+  PING_LENGTH: 32
+}
