+++ src/util.js
@@ -4,6 +4,7 @@ const randomBytes = require('random-bytes')
 const config = require('./config')
 
 exports = module.exports
+
 exports.genPingValue = (length) => {
   if (!length) {
     length = config.PING_LENGTH
